TUI: checkpoint with add form working well
