+++ src/libro/tui/utils.py
@@ -1,0 +1,4 @@
+"""Utility functions for the TUI interface"""
+
+# This file is reserved for shared utilities across TUI components
+# Currently empty but provides structure for future utilities
